private load, split tests
